pyproject.toml
[tool.poetry]
name = "flask-chatgpt-app"
version = "0.1.0"
description = "A simple Flask app to interact with ChatGPT"
authors = ["Your Name <youremail@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
flask = "^2.0.0"
openai = "*"
python-dotenv = "^0.21.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"